apps/dlq_handler/requirements.txt
kafka-python==2.0.2
elasticsearch==8.13.1
python-json-logger==2.0.7
requests==2.32.3
prometheus-client==0.20.0
